:art: Format code and optimize imports

+++ app/src/main/java/com/kyhsgeekcode/filechooser/model/FileItemDotNetSymbol.kt
@@ -42,7 +42,7 @@ fun getValueFromTypeKindAndBytes(bytes: ByteArray, kind: Int): Any {
     val bb = ByteBuffer.wrap(bytes).order(ByteOrder.LITTLE_ENDIAN)
     return when (kind) {
         TypeKind.ELEMENT_TYPE_BOOLEAN -> bytes[0].toInt() != 0
-        TypeKind.ELEMENT_TYPE_CHAR -> bytes[0].toChar()
+        TypeKind.ELEMENT_TYPE_CHAR -> bytes[0].toInt().toChar()
         TypeKind.ELEMENT_TYPE_I -> bb.int
         TypeKind.ELEMENT_TYPE_I1 -> bb.get()
         TypeKind.ELEMENT_TYPE_I2 -> bb.short
